fix: refactor/relocate for review comments
